Adds utils.py with the collision detection function in it
